Remove the cleanup_instance routine
Reverted parts of https://github.com/mitchellh/vagrant-google/pull/113
due to dangerous behavior.
If an instance of the same name already exists in the zone it would
catch the Fog::Errors::Error that instance already exists and would
proceed to clean it up, destroying the instance and leaving a disk
behind.
The behavior it was designed to catch should be a serious backend bug
anyway and needs to be reported.