fix: Set a default zone only if `default` network is used
Previously, if the zone was not set in a configuration then we were
always setting the zone to `us-central1-f` as a default. This zone is a
reasonable default when we expect that the `default` network is being
used. However, using this as the automatic default in the case that a
non-default network was set but not a zone throws an error that is
misleading.

This change adds some logic to test if the `default` network is being
used before setting the zone to the default. This should cause a more
useful error to surface when the network is set but the zone isn't.

Unit tests included.